feat: add language switcher component
