+++ src/components/Navigation.tsx
@@ -1,24 +1,24 @@
 import React, { useState, useEffect } from "react";
 import { ReactComponent as TolocarLogoSvg } from "@assets/tolocar_logo.svg";
 import { ReactComponent as ArrowIcon } from "@assets/arrow.svg";
-import LanguageUtils from "@util/LanguageUtils";
 import type { IMenuItem } from "@interfaces/IMenu";
+import { LanguageSwitcher } from "@components";
 
 interface Props {
   className?: string;
   baseUrl?: string;
   path?: string;
-  menu?: IMenuItem[];
   locale: string;
+  menu?: IMenuItem[];
   dark?: boolean;
 }
 
 const Navigation: React.FC<Props> = ({
   menu,
   baseUrl = "/",
-  className,
   path,
   locale,
+  className,
   dark,
 }: Props) => {
   const [showOverlayMenu, setShowOverlayMenu] = useState(false);
@@ -40,21 +40,6 @@ const Navigation: React.FC<Props> = ({
 
   const toggleMenu = () => {
     setShowOverlayMenu(!showOverlayMenu);
-  };
-
-  // Changing the language requires only to change the locale in the URL
-  // The rest (slug) will be handled by redirect logic
-  const changeLanguage = (newLanguage: string) => {
-    LanguageUtils.setLanguage(newLanguage);
-    const pathWithoutBaseUrl = path
-      ?.replace(baseUrl, "")
-      ?.split("/")
-      .filter(Boolean)
-      .slice(1)
-      .join("/");
-    const newLocation = `${baseUrl}${newLanguage}/${pathWithoutBaseUrl}`;
-
-    window.location.pathname = newLocation;
   };
 
   const Image = (
@@ -93,48 +78,20 @@ const Navigation: React.FC<Props> = ({
                   </MenuListItem>
                 )
             )}
-            <ul className="flex items-center gap-2 lg:pl-[38px]">
-              <LanguageSwitcherItem
-                onClick={() => {
-                  changeLanguage("ua");
-                }}
-                isSelected={locale === "ua"}
-              >
-                UA
-              </LanguageSwitcherItem>
-              <span className="text-base">/</span>
-              <LanguageSwitcherItem
-                onClick={() => {
-                  changeLanguage("en");
-                }}
-                isSelected={locale === "en"}
-              >
-                EN
-              </LanguageSwitcherItem>
-            </ul>
+            <LanguageSwitcher
+              path={path}
+              locale={locale}
+              className="gap-2 lg:pl-[38px]"
+            />
           </ul>
         </nav>
         <div className="flex items-center md:hidden bg-white box-border z-20">
           {showOverlayMenu && (
-            <ul className="flex items-center gap-3 text-neutral-500 font-medium mr-4">
-              <LanguageSwitcherItem
-                onClick={() => {
-                  changeLanguage("ua");
-                }}
-                isSelected={locale === "ua"}
-              >
-                UA
-              </LanguageSwitcherItem>
-              <span className="text-base">/</span>
-              <LanguageSwitcherItem
-                onClick={() => {
-                  changeLanguage("en");
-                }}
-                isSelected={locale === "en"}
-              >
-                EN
-              </LanguageSwitcherItem>
-            </ul>
+            <LanguageSwitcher
+              path={path}
+              locale={locale}
+              className="gap-3 text-neutral-500 font-medium mr-4"
+            />
           )}
 
           <HamburgerButton onClick={toggleMenu} isOpen={showOverlayMenu} />
@@ -264,38 +221,6 @@ const OverlayMenu: React.FC<OverlayMenuProps> = ({
         </div>
       </div>
     </div>
-  );
-};
-
-interface LanguageSwitcherItemProps {
-  className?: string;
-  onClick?: () => void;
-  children: React.ReactNode;
-  isSelected?: boolean;
-}
-const LanguageSwitcherItem: React.FC<LanguageSwitcherItemProps> = ({
-  className,
-  onClick,
-  children,
-  isSelected,
-}) => {
-  return (
-    <li
-      className={`flex-shrink-0 md:rounded-md overflow-hidden ${
-        className || ""
-      }`}
-    >
-      <a
-        onClick={isSelected ? undefined : onClick}
-        className={`md:hover:text-neutral-800 ${
-          isSelected ? "text-neutral-800 " : "cursor-pointer "
-        }`}
-      >
-        <div className="flex text-lg lg:text-[15px] font-inter items-center md:text-base md:font-medium">
-          {children}
-        </div>
-      </a>
-    </li>
   );
 };
